Fixed some warnings on Windows
Windows VC compiler complains about comparison between signed and unsigned
int types, unused variables


git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@25205 dc483132-0cff-0310-8789-dd5450dbe970

+++ src/appl/gss-sample/gss-client.c
@@ -395,7 +395,7 @@ read_file(file_name, in_buf)
         perror("read");
         exit(1);
     }
-    if (count < in_buf->length)
+    if (count < (int)in_buf->length)
         fprintf(stderr, "Warning, only read in %d bytes, expected %d\n",
                 count, (int) in_buf->length);
 }
@@ -563,7 +563,7 @@ call_server(host, port, oid, service_name, gss_flags, auth_flag,
         in_buf.length = strlen((char *)in_buf.value);
     }
 
-    for (i = 0; i < mcount; i++) {
+    for (i = 0; i < (size_t)mcount; i++) {
         if (wrap_flag) {
             maj_stat =
                 gss_wrap(&min_stat, context, encrypt_flag, GSS_C_QOP_DEFAULT,
